add test_helpers containing build_dataset
